V1.4.2

+++ tasks/maintenance/module.py
@@ -10,7 +10,7 @@ from core.utils.file import File
 from core.utils.helpers import check_status, prepare_str, check_edit_age
 from core.utils.pipeline import Pipeline
 from core.utils.wikidb import Database
-from database.models import Page, Status, TaskName
+from database.models import Page, Status as Model_Status
 
 from tasks.maintenance.bots.dead_end import DeadEnd
 from tasks.maintenance.bots.has_categories import HasCategories
@@ -76,14 +76,23 @@ FROM (
         yield title
 
 
-def get_skip_pages():
+# todo: call it from bot task
+def get_skip_pages(name_of_page = None):
     home_path = os.path.expanduser("~")
     file = File(script_dir=home_path)
     file_path = prepare_str('maintenance_skip.txt')
     file.set_stub_path(file_path)
     file.get_json_content()
     templates = json.loads(file.contents)
-    return templates
+    if name_of_page is None:
+        return templates
+    else:
+        found = False
+        for page in templates:
+            if prepare_str(page) == prepare_str(name_of_page):
+                found = True
+                break
+        return  found
 
 
 class PipelineTasks:
@@ -151,45 +160,46 @@ class ProcessArticle:
             # get page object
             self.page = pywikibot.Page(self.site, self.title)
             # Check if the page has already been processed
-            self.page_query = self.session.query(Page).filter_by(id=self.id, status=Status.PENDING).one_or_none()
+            self.page_query = self.session.query(Page).filter_by(id=self.id, status=Model_Status.PENDING).one_or_none()
 
-            # Update the status of the page to indicate that it is being processed
-            self.page_query.status = Status.RECEIVED
-            self.session.commit()
 
             if self.page is None:
                 self._delete_page()
             else:
-                if self.page.exists() and (not self.page.isRedirectPage()):
-                    # if status true can edit
-                    if check_edit_age(page=self.page):
+                if self.page_query is not None:
+                    # Update the status of the page to indicate that it is being processed
+                    self.page_query.status = Model_Status.RECEIVED
+                    self.session.commit()
 
-                        try:
+                    if self.page.exists() and (not self.page.isRedirectPage()):
+                        # if status true can edit
+                        if check_edit_age(page=self.page) and not get_skip_pages(name_of_page=self.page.title(with_ns=False)):
+                            try:
 
-                            self.pipeline = Pipeline(self.page, self.page.text, TASK_SUMMARY, PipelineTasks.steps,
-                                                     PipelineTasks.extra_steps)
-                            processed_text, processed_summary = self.pipeline.process()
-                            # write processed text back to the page
-                            if self.pipeline.hasChange() and check_status("مستخدم:LokasBot/إيقاف مهمة صيانة المقالات"):
-                                print("start save " + self.page.title())
-                                self.page.text = processed_text
-                                self.page.save(summary=clean_summary(processed_summary))
-                            else:
-                                print("page not changed " + self.page.title())
+                                self.pipeline = Pipeline(self.page, self.page.text, TASK_SUMMARY, PipelineTasks.steps,
+                                                         PipelineTasks.extra_steps)
+                                processed_text, processed_summary = self.pipeline.process()
+                                # write processed text back to the page
+                                if self.pipeline.hasChange() and check_status("مستخدم:LokasBot/إيقاف مهمة صيانة المقالات"):
+                                    print("start save " + self.page.title())
+                                    self.page.text = processed_text
+                                    self.page.save(summary=clean_summary(processed_summary))
+                                else:
+                                    print("page not changed " + self.page.title())
 
-                            self._delete_page()
+                                self._delete_page()
 
-                        except Exception as e:
-                            logging.error(f"An error occurred while processing {self.title}: {e}")
-                            logging.exception(e)
-                            if self.page_query is not None:
-                                self._delay_page(hours=1)
+                            except Exception as e:
+                                logging.error(f"An error occurred while processing {self.title}: {e}")
+                                logging.exception(e)
+                                if self.page_query is not None:
+                                    self._delay_page(hours=1)
+
+                        else:
+                            self._delay_page(hours=1)
 
                     else:
-                        self._delay_page(hours=1)
-
-                else:
-                    self._delete_page()
+                        self._delete_page()
 
 
         except Exception as e:
@@ -199,15 +209,16 @@ class ProcessArticle:
                 self._delay_page(hours=1)
 
     def _delete_page(self):
-        # Delete the page from the database
-        self.session.delete(self.page_query)
-        self.session.commit()
+        if self.page_query is not None:
+            # Delete the page from the database
+            self.session.delete(self.page_query)
+            self.session.commit()
 
     def _delay_page(self, hours=1):
         # Update the status of the page to indicate that it needs to be processed again later
-
-        delta = datetime.timedelta(hours=hours)
-        new_date = datetime.datetime.now() + delta
-        self.page_query.status = Status.PENDING
-        self.page_query.date = new_date
-        self.session.commit()
+        if self.page_query is not None:
+            delta = datetime.timedelta(hours=hours)
+            new_date = datetime.datetime.now() + delta
+            self.page_query.status = Model_Status.PENDING
+            self.page_query.update_date = new_date
+            self.session.commit()
